pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.84.0",
    "extra-streamlit-components>=0.1.81",
    "google-genai>=1.67.0",
    "langchain-anthropic>=1.3.5",
    "langchain-core>=1.2.19",
    "langchain-google-genai>=4.2.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sift-stack-py>=0.12.0",
    "streamlit>=1.55.0",
    "streamlit-authenticator>=0.4.2",
]
